move web error code enum to web/enums
