+++ lib/src/web/enums/web_error_code.dart
